requirements.txt
python-telegram-bot

langchain
